+++ gulpfile.js
@@ -5,6 +5,8 @@ var sass = require('gulp-sass');
 var server = require('gulp-server-livereload');
 var normalize = require('node-normalize-scss').includePaths;
 var autoprefixer = require('gulp-autoprefixer');
+var uglify = require('gulp-uglify');
+var concat = require('gulp-concat');
 
 gulp.task('webserver', ['watch'], function() {
   gulp.src('./')
@@ -38,5 +40,12 @@ gulp.task('prefixer', function () {
     .pipe(gulp.dest('assets/css-min'))
 });
 
-gulp.task('default', [ 'sass', 'webserver' ]);
+gulp.task('js-uglify', function (cb) {
+  gulp.src('assets/js/*.js')
+    .pipe(concat('app.js'))
+    .pipe(uglify())
+    .pipe(gulp.dest('assets/js-min/'))
+});
+
+gulp.task('default', [ 'sass', 'js-uglify', 'webserver' ]);
 gulp.task('watch', [ 'sass:watch' ]);
