+++ index.html
@@ -7,7 +7,6 @@
     <meta name="viewport" content="initial-scale=1.0, maximum-scale=1.0, user-scalable=no"/>
     <meta name="theme-color" content="#EA5653" />
     <link rel="stylesheet" media="all" href="assets/css-min/app.css" />
-
     <script src="https://code.jquery.com/jquery-1.11.3.min.js"></script>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery-mockjax/1.6.2/jquery.mockjax.min.js"></script>
     <script>$.mockjax({ url: /.*\/api\/newsletter/, responseText: { status: 'success', message: 'email cadastrado com sucesso' }});</script>
@@ -84,7 +83,6 @@
         </div>
       </div>
     </section>
-
-    <script src="assets/js/app.js"></script>
+    <script src="assets/js-min/app.js"></script>
   </body>
 </html>
