+++ package.json
@@ -17,8 +17,10 @@
   "devDependencies": {
     "gulp": "^3.9.1",
     "gulp-autoprefixer": "^3.1.1",
+    "gulp-concat": "^2.6.1",
     "gulp-sass": "^3.1.0",
     "gulp-server-livereload": "^1.9.2",
+    "gulp-uglify": "^2.0.1",
     "node-normalize-scss": "^1.4.0"
   }
 }
